Houdini CSV Asset: - Packed vector values in the CSV file are now properly supported (not just for Position/Normal) and can be of any size. - Added support for reimporting Houdini CSV files. - Added an "open in text editor" entry in the context menu. - Improved error/warning logging during the parsing of the file
Houdini Niagara Data Interface:
- Added GetParticleIndexesToSpawnAtTime():
New helper functions returning the min index, max index and number of particles to be spawned for a given time value.
Uses an internal LastSpawnIndex to avoid spawning the same particles twice.
- Modified GetLastParticleIndexAtTime():
If the CSV file doesn't have time informations, returns false and set the LastIndex to the last particle
If desiredTime is smaller than the first particle, LastIndex will be set to -1
If desiredTime is higher than the last particle in the csv file, LastIndex will be set to the last particle's index